callbacks don't submit futures
